+++ packages/polywrap-core/polywrap_core/types/__init__.py
@@ -1,4 +1,4 @@
-from .client import *
+from .core_client import *
 from .env import *
 from .file_reader import *
 from .invoke import *
