+++ packages/polywrap-core/polywrap_core/types/core_client.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 from abc import abstractmethod
 from dataclasses import dataclass, field
 from typing import List, Optional, Union, Dict, Any
+from .interface_implementation import InterfaceImplementations
 
 from polywrap_manifest import AnyWrapManifest, DeserializeManifestOptions
 from polywrap_result import Result
